fix(fileutils): added logic to handle non-existing path's body
Previously, if the paths body was not available, and exception was
raised causing the operation to stop in the middle. The worst part is,
`finally` block of the `makefiles.mkfile.main` function caused this
exception to be silent. This commit removed that `finally` block, and
added checks for non-existing body of path.